+++ app/src/main/java/com/westas/orderassembly/invoice_items/TOnChangeQuantity.java
@@ -3,5 +3,5 @@ package com.westas.orderassembly.invoice_items;
 import com.westas.orderassembly.rest_service.TResponce;
 
 public interface TOnChangeQuantity {
-    void EventChangeQuantity(String uid, String barcode, double quantity);
+    void EventChangeQuantity(String uid, String barcode, float quantity);
 }
